refactor(coinbase): centralize types and enforce strict typing
- Create central types.ts file for common blockchain types
- Define Address, PrivateKey, TransactionHash as hex string types
- Move FaucetNetwork, FaucetToken, FaucetConfig to types.ts
- Enforce Secret<PrivateKey> for better type safety
- Update all usages to use centralized types

This provides stronger type safety and prevents repetition across the codebase.
