+++ alchemy/src/coinbase/index.ts
@@ -1,2 +1,3 @@
+export * from "./types.ts";
 export * from "./evm-account.ts";
 export * from "./evm-smart-account.ts";
